move bot code into folder and put into a class for code consistency
